chore: 컴포넌트 내 type 파일로 분리
