+++ src/components/organisms/BlockedUserItem/index.tsx
@@ -1,17 +1,8 @@
 import { TextButton } from "components/atoms";
 import { Profile } from "../Profile";
 import { BlockedUserItemWrapper } from "./styled";
+import type { IProfile } from "types";
 
-export interface IProfile {
-  /** 프로필 이미지 URL */
-  imgUrl: string;
-  /** 닉네임 */
-  nickname: string;
-  /** 인증한 지역 */
-  location: string;
-  /** 차단 여부 */
-  isBlocked: boolean;
-}
 interface IBlockedUserItemWrapperProps {
   /** 유저 프로필 */
   profile: IProfile;
