+++ src/components/organisms/BlockedUsersList/index.tsx
@@ -1,14 +1,10 @@
-import { IProfile, BlockedUserItem } from "../BlockedUserItem";
+import { BlockedUserItem } from "../BlockedUserItem";
 import { BlockedUsersListWrapper } from "./styled";
+import type { IBlockedUserItem } from "types";
 
 interface IBlockedUsersListProps {
   /** BlockedUserItem 리스트 */
   blockedUserItems: IBlockedUserItem[];
-}
-
-export interface IBlockedUserItem {
-  profile: IProfile;
-  onClick: () => void;
 }
 
 export const BlockedUsersList = ({
@@ -21,7 +17,7 @@ export const BlockedUsersList = ({
           <BlockedUserItem
             key={idx}
             profile={blockedUserItem.profile}
-            onClick={blockedUserItem.onClick}
+            onClick={() => blockedUserItem.onClick(idx)}
           ></BlockedUserItem>
         );
       })}
